chore(web): split operator-menu dropdown into its own file (ROK-1323)
